+++ packages/editor/src/editor-context.ts
@@ -1,5 +1,5 @@
 import { JsonSchema } from '@jsonforms/core';
-import { ModelMapping } from './helpers/containment.util';
+import { ModelMapping } from './helpers/container.util';
 
 // TODO remove when it is no longer needed by the schema service impl: Besides that it is obsolete
 export interface EditorContext {
